backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pdfplumber>=0.10.2
pdf2image>=1.16.3
pytesseract>=0.3.10
python-docx>=0.8.11
pypandoc>=1.11
httpx>=0.24.0
numpy>=1.24.0
Pillow>=10.0.0
python-dotenv>=1.0.0
huggingface_hub>=0.19.0
groq>=0.4.2
razorpay>=1.4.1
supabase>=2.3.0
